refactor(shortform): split export status support
